Exclude Anchor Protocol Market contract
The Anchor protocol is a dead protocol with no active development and smart contracts that are not compatible with the current version of wasmvm used with Terra Classic. It is also an "insider" protocol as per the CMC definition of circulating supply. For these reasons it is now excluded from the circulating supply calculation.